+++ iOS/LeafByte/Utils/UIUtils.swift
@@ -37,7 +37,7 @@ func finishWithImagePicker(self viewController: UIViewController, info: [UIImage
         }
     } else {
         onDismissingPicker = {
-            presentAlert(self: viewController, title: "Failed", message: "Failed to open chosen image. Please reach out to leafbyte@zoegp.science with information about what image you chose so we can fix this issue. Debug info: \(info)")
+            presentAlert(self: viewController, title: nil, message: "Failed to open chosen image. Please reach out to leafbyte@zoegp.science with information about what image you chose so we can fix this issue. Debug info: \(info)")
         }
     }